refactor!: replace `Spec` type with `graph` pkg, introduce interfaces
This commit refactors a large part of the repository. The `Spec` type is replaced by the `graph` package and most of the types are now `interfaces`, allowing to provide custom data in future features.

Most of the templates are also changed, improving readability.

Other changes:
- Replace `Generator.Generate` with `Generator.GenerateSchema` and `Generator.GenerateGraph`

BREAKING CHANGE:
Need to update the `Generator` calls and template execution context